Remove "size" option to vmplot.
This makes usage simpler, as we don't have to manually run wc -l on the
file before we call vmplot.
This makes usage more reliable, as we don't have to be careful with
filenames and arguments.  (Leaving out the -s option previously meant it
ignored your specified file argument)

This DOES makes vmplot use more memory, as the entire log file is read into
memory before being written out again, instead of writing it line by
line as it was read.  Even an entire day's worth log of logging at 1 second
intervals is only 86k lines though, so this shouldn't be any problem in
practice.